Refactor legacy loginHint lookup
